add data

+++ data/src/main/java/com/example/data/models/Date.kt
@@ -1,6 +1,6 @@
 package com.example.data.models
 
 data class Date(
-    val end: Int,
-    val start: Int
+    val end: Long,
+    val start: Long
 )
